backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langgraph
langchain-groq
pydantic
python-dotenv
langchain-core
langchain-community
